Specify types for partitioning columns.

+++ src/hipscat_import/index/map_reduce.py
@@ -27,9 +27,9 @@ def create_index(args):
 
     if args.include_order_pixel:
         ## Take out the hive dictionary behavior.
-        data["Norder"] = data["Norder"].astype(np.int32)
-        data["Dir"] = data["Dir"].astype(np.int32)
-        data["Npix"] = data["Npix"].astype(np.int32)
+        data["Norder"] = data["Norder"].astype(np.uint8)
+        data["Dir"] = data["Dir"].astype(np.uint32)
+        data["Npix"] = data["Npix"].astype(np.uint32)
     data = data.reset_index()
     if not args.include_hipscat_index:
         data = data.drop(columns=[HIPSCAT_ID_COLUMN])
